requirements_api.txt
fastapi[standard,docs,testing]
pydantic-settings
fastapi-cache2
pyjwt
